requirements.txt
# Core Dependencies
langchain==0.3.13
langchain-community==0.4.1
langchain-core==1.1.3
langchain-text-splitters==1.0.0
langchain-ollama==0.3.0

# Vector Store
chromadb==0.6.3

# Code Parsing
tree-sitter==0.25.2
tree-sitter-cpp==0.25.1

# Visualization
graphviz==0.20.3
pillow==11.1.0

# Utilities
python-dotenv==1.0.1
pyyaml==6.0.2
rich==13.9.4
tqdm==4.67.1

